Devanagiri-OCR/requirements.txt
tensorflow-gpu==2.8.0
sklearn==0.0
pandas==1.4.2
matplotlib==3.5.1